Introduce mixins for the tests. Use mixins in the existing tests. Also, a minor cleanup.
