+++ functional-tests/tests/crash_duty_sign_block.py
@@ -1,12 +1,21 @@
 import flexitest
 
-from envs import testenv
+from mixins import seq_crash_mixin
+from utils import wait_until
 
 
 @flexitest.register
-class CrashDutySignBlockTest(testenv.CrashTestBase):
+class CrashDutySignBlockTest(seq_crash_mixin.SeqCrashMixin):
     def __init__(self, ctx: flexitest.InitContext):
-        super().__init__(ctx)
+        ctx.set_env("basic")
 
-    def get_bail_tag(self) -> str:
-        return "duty_sign_block"
+    def main(self, ctx: flexitest.RunContext):
+        cur_chain_tip = self.handle_bail(lambda: "duty_sign_block")
+
+        wait_until(
+            lambda: self.seqrpc.strata_clientStatus()["chain_tip_slot"] > cur_chain_tip,
+            error_with="chain tip slot not progressing",
+            timeout=20,
+        )
+
+        return True
